+++ zcash_client_backend/CHANGELOG.md
@@ -22,6 +22,7 @@ workspace.
     - `standalone`
     - `exposure`
     - `next_check_time`
+- `zcash_client_backend::data_api::WalletUtxo`
 - `zcash_client_backend::data_api::WalletTest::get_known_ephemeral_addresses`
 - `zcash_client_backend::data_api::WalletTest::find_account_for_ephemeral_address`
 
@@ -34,6 +35,11 @@ workspace.
   - `InputSource::{get_spendable_note, get_account_metadata, get_unspent_transparent_output}`
     each now take an additional `target_height` argument; spendability isn't a
     well-defined property in absence of target height information.
+  - The result types of `InputSource::get_unspent_transparent_output` and
+    `InputSource::get_unspent_transparent_outputs` have each changed; instead
+    of returning bare `WalletTransparentOutput`s, these now return `WalletUtxo`
+    values that include the `WalletTransparentOutput` data along with additional
+    derivation metadata.
   - `WalletRead` has added method `get_ephemeral_transparent_receivers`.
   - The result type of `WalletRead::get_transparent_receivers` has changed. The
     value type of the returned `HashMap` is now non-optional.
@@ -46,6 +52,11 @@ workspace.
     details.
 - `zcash_client_backend::fees::ChangeStrategy::fetch_wallet_meta` now takes
   an additional `target_height` argument.
+- Variants of `zcash_client_backend::proposal::ProposalError` have changed.
+  A new `EphemeralAddressLinkability` variant has been added, to represent
+  the case where a caller attempts to construct a shielding transaction that
+  would link an ephemeral address to any other transparent address in the wallet
+  on-chain.
 - `zcash_client_backend::wallet`:
   - `TransparentAddressMetadata` has been converted from an enum to a struct
     that contains both source metadata and information about when the address